agent_evalution/requirements.txt
langchain==1.2.7
langchain_aws==1.2.1
langfuse==3.12.0
python-dotenv==1.2.1
pydantic==2.12.5
agentevals==0.0.9
guardrails-ai==0.7.2